refactor: split up Util class
