+++ src/XIVLauncher.Common.Unix/Compatibility/GameFixes/Implementations/MacVideoFix.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Net.Http;
 using System.Threading.Tasks;
+using XIVLauncher.Common.Util;
 
 namespace XIVLauncher.Common.Unix.Compatibility.GameFixes.Implementations;
 
@@ -31,7 +32,7 @@ public class MacVideoFix : GameFix
         {
             if (percentage != null && size != null)
             {
-                this.UpdateProgress?.Invoke($"{LoadingTitle} ({Util.BytesToString(downloaded)}/{Util.BytesToString(size.Value)})", true, (float)(percentage.Value / 100f));
+                this.UpdateProgress?.Invoke($"{LoadingTitle} ({ApiHelpers.BytesToString(downloaded)}/{ApiHelpers.BytesToString(size.Value)})", true, (float)(percentage.Value / 100f));
             }
         };
 
@@ -58,88 +59,5 @@ public class MacVideoFix : GameFix
 
         Directory.Delete(tempMacExtract, true);
         File.Delete(zipFilePath);
-    }
-
-    private class HttpClientDownloadWithProgress : IDisposable
-    {
-        private readonly string downloadUrl;
-        private readonly string destinationFilePath;
-
-        private HttpClient httpClient;
-
-        public delegate void ProgressChangedHandler(long? totalFileSize, long totalBytesDownloaded, double? progressPercentage);
-
-        public event ProgressChangedHandler ProgressChanged;
-
-        public HttpClientDownloadWithProgress(string downloadUrl, string destinationFilePath)
-        {
-            this.downloadUrl = downloadUrl;
-            this.destinationFilePath = destinationFilePath;
-        }
-
-        public async Task StartDownload()
-        {
-            this.httpClient = new HttpClient { Timeout = TimeSpan.FromDays(1) };
-
-            using var response = await this.httpClient.GetAsync(this.downloadUrl, HttpCompletionOption.ResponseHeadersRead).ConfigureAwait(false);
-            await this.DownloadFileFromHttpResponseMessage(response).ConfigureAwait(false);
-        }
-
-        private async Task DownloadFileFromHttpResponseMessage(HttpResponseMessage response)
-        {
-            response.EnsureSuccessStatusCode();
-
-            var totalBytes = response.Content.Headers.ContentLength;
-
-            using var contentStream = await response.Content.ReadAsStreamAsync().ConfigureAwait(false);
-            await this.ProcessContentStream(totalBytes, contentStream).ConfigureAwait(false);
-        }
-
-        private async Task ProcessContentStream(long? totalDownloadSize, Stream contentStream)
-        {
-            var totalBytesRead = 0L;
-            var readCount = 0L;
-            var buffer = new byte[8192];
-            var isMoreToRead = true;
-
-            using var fileStream = new FileStream(this.destinationFilePath, FileMode.Create, FileAccess.Write, FileShare.None, 8192, true);
-
-            do
-            {
-                var bytesRead = await contentStream.ReadAsync(buffer, 0, buffer.Length).ConfigureAwait(false);
-
-                if (bytesRead == 0)
-                {
-                    isMoreToRead = false;
-                    this.TriggerProgressChanged(totalDownloadSize, totalBytesRead);
-                    continue;
-                }
-
-                await fileStream.WriteAsync(buffer, 0, bytesRead).ConfigureAwait(false);
-
-                totalBytesRead += bytesRead;
-                readCount += 1;
-
-                if (readCount % 100 == 0)
-                    this.TriggerProgressChanged(totalDownloadSize, totalBytesRead);
-            } while (isMoreToRead);
-        }
-
-        private void TriggerProgressChanged(long? totalDownloadSize, long totalBytesRead)
-        {
-            if (this.ProgressChanged == null)
-                return;
-
-            double? progressPercentage = null;
-            if (totalDownloadSize.HasValue)
-                progressPercentage = Math.Round((double)totalBytesRead / totalDownloadSize.Value * 100, 2);
-
-            this.ProgressChanged(totalDownloadSize, totalBytesRead, progressPercentage);
-        }
-
-        public void Dispose()
-        {
-            this.httpClient?.Dispose();
-        }
     }
 }
