+++ src/XIVLauncher.Common/Game/Headlines.cs
@@ -4,6 +4,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+using XIVLauncher.Common.Util;
 
 namespace XIVLauncher.Common.Game
 {
@@ -53,7 +54,7 @@ namespace XIVLauncher.Common.Game
     {
         public static async Task<Headlines> Get(Launcher game, ClientLanguage language)
         {
-            var unixTimestamp = Util.GetUnixMillis();
+            var unixTimestamp = ApiHelpers.GetUnixMillis();
             var langCode = language.GetLangCode();
             var url = $"https://frontier.ffxiv.com/news/headline.json?lang={langCode}&media=pcapp&_={unixTimestamp}";
