+++ src/XIVLauncher.Common/PlatformAbstractions/IDalamudRunner.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using XIVLauncher.Common.Dalamud;
