+++ src/XIVLauncher.Core/Components/MainPage/MainPage.cs
@@ -15,6 +15,7 @@ using XIVLauncher.Common.Windows;
 using XIVLauncher.Common.Unix;
 using XIVLauncher.Common.Unix.Compatibility;
 using XIVLauncher.Common.Unix.Compatibility.GameFixes;
+using XIVLauncher.Common.Util;
 using XIVLauncher.Core.Accounts;
 
 namespace XIVLauncher.Core.Components.MainPage;
@@ -47,7 +48,7 @@ public class MainPage : Page
 
         if (savedAccount != null) this.SwitchAccount(savedAccount, false);
 
-        if (Util.IsElevated())
+        if (PlatformHelpers.IsElevated())
             App.ShowMessage("XIVLauncher is running as administrator/root user.\nThis can cause various issues, including but not limited to addons failing to launch and hotkey applications failing to respond.\n\nPlease take care to avoid running XIVLauncher with elevated privileges", "XIVLauncher");
     }
 
@@ -925,7 +926,7 @@ public class MainPage : Page
         }
 #endif
 
-        if (Util.CheckIsGameOpen())
+        if (GameHelpers.CheckIsGameOpen())
         {
             App.ShowMessageBlocking(
                 Loc.Localize("GameIsOpenError",
@@ -971,8 +972,8 @@ public class MainPage : Page
                     Thread.Sleep(30);
 
                     App.LoadingPage.Line2 = string.Format("Working on {0}/{1}", patcher.CurrentInstallIndex, patcher.Downloads.Count);
-                    App.LoadingPage.Line3 = string.Format("{0} left to download at {1}/s", Util.BytesToString(patcher.AllDownloadsLength < 0 ? 0 : patcher.AllDownloadsLength),
-                        Util.BytesToString(patcher.Speeds.Sum()));
+                    App.LoadingPage.Line3 = string.Format("{0} left to download at {1}/s", ApiHelpers.BytesToString(patcher.AllDownloadsLength < 0 ? 0 : patcher.AllDownloadsLength),
+                        ApiHelpers.BytesToString(patcher.Speeds.Sum()));
 
                     App.LoadingPage.Progress = patcher.CurrentInstallIndex / (float)patcher.Downloads.Count;
                 }
@@ -1007,7 +1008,7 @@ public class MainPage : Page
                         string.Format(
                             Loc.Localize("FreeSpaceError",
                                 "There is not enough space on your drive to download patches.\n\nYou can change the location patches are downloaded to in the settings.\n\nRequired:{0}\nFree:{1}"),
-                            Util.BytesToString(sex.BytesRequired), Util.BytesToString(sex.BytesFree)), "XIVLauncher Error");
+                            ApiHelpers.BytesToString(sex.BytesRequired), ApiHelpers.BytesToString(sex.BytesFree)), "XIVLauncher Error");
                     break;
 
                 case NotEnoughSpaceException.SpaceKind.AllPatches:
@@ -1015,7 +1016,7 @@ public class MainPage : Page
                         string.Format(
                             Loc.Localize("FreeSpaceErrorAll",
                                 "There is not enough space on your drive to download all patches.\n\nYou can change the location patches are downloaded to in the XIVLauncher settings.\n\nRequired:{0}\nFree:{1}"),
-                            Util.BytesToString(sex.BytesRequired), Util.BytesToString(sex.BytesFree)), "XIVLauncher Error");
+                            ApiHelpers.BytesToString(sex.BytesRequired), ApiHelpers.BytesToString(sex.BytesFree)), "XIVLauncher Error");
                     break;
 
                 case NotEnoughSpaceException.SpaceKind.Game:
@@ -1023,7 +1024,7 @@ public class MainPage : Page
                         string.Format(
                             Loc.Localize("FreeSpaceGameError",
                                 "There is not enough space on your drive to install patches.\n\nYou can change the location the game is installed to in the settings.\n\nRequired:{0}\nFree:{1}"),
-                            Util.BytesToString(sex.BytesRequired), Util.BytesToString(sex.BytesFree)), "XIVLauncher Error");
+                            ApiHelpers.BytesToString(sex.BytesRequired), ApiHelpers.BytesToString(sex.BytesFree)), "XIVLauncher Error");
                     break;
 
                 default:
