+++ src/XIVLauncher/AppUtil.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Reflection;
 using Microsoft.Win32;
 using XIVLauncher.Common;
+using XIVLauncher.Common.Util;
 
 namespace XIVLauncher
 {
@@ -96,7 +97,7 @@ namespace XIVLauncher
 
                 foreach (var path in GetCommonPaths())
                 {
-                    if (!Directory.Exists(path) || !Util.IsValidFfxivPath(path) || foundVersions.ContainsKey(path))
+                    if (!Directory.Exists(path) || !GameHelpers.IsValidFfxivPath(path) || foundVersions.ContainsKey(path))
                         continue;
 
                     var baseVersion = Repository.Ffxiv.GetVer(new DirectoryInfo(path));
@@ -115,7 +116,7 @@ namespace XIVLauncher
                                 // DisplayIcon includes "boot\ffxivboot.exe", need to remove it
                                 path = Directory.GetParent(path).Parent.FullName;
 
-                                if (Directory.Exists(path) && Util.IsValidFfxivPath(path) && !foundVersions.ContainsKey(path))
+                                if (Directory.Exists(path) && GameHelpers.IsValidFfxivPath(path) && !foundVersions.ContainsKey(path))
                                 {
                                     var baseVersion = Repository.Ffxiv.GetVer(new DirectoryInfo(path));
                                     foundVersions.Add(path, SeVersion.Parse(baseVersion));
@@ -130,7 +131,7 @@ namespace XIVLauncher
                             {
                                 if (subkey != null && subkey.GetValue("InstallLocation", null) is string path)
                                 {
-                                    if (Directory.Exists(path) && Util.IsValidFfxivPath(path) && !foundVersions.ContainsKey(path))
+                                    if (Directory.Exists(path) && GameHelpers.IsValidFfxivPath(path) && !foundVersions.ContainsKey(path))
                                     {
                                         // InstallLocation is the root path of the game (the one containing boot and game) itself
                                         var baseVersion = Repository.Ffxiv.GetVer(new DirectoryInfo(path));
@@ -139,7 +140,6 @@ namespace XIVLauncher
                                 }
                             }
                         }
-
                     }
                 }
