+++ src/XIVLauncher/Windows/FirstTimeSetupWindow.xaml.cs
@@ -7,6 +7,7 @@ using IWshRuntimeLibrary;
 using XIVLauncher.Common;
 using XIVLauncher.Common.Addon;
 using XIVLauncher.Common.Addon.Implementations;
+using XIVLauncher.Common.Util;
 using XIVLauncher.Windows.ViewModel;
 
 namespace XIVLauncher.Windows
@@ -80,26 +81,26 @@ namespace XIVLauncher.Windows
                     return;
                 }
 
-                if (!Util.LetChoosePath(GamePathEntry.Text))
+                if (!GameHelpers.LetChoosePath(GamePathEntry.Text))
                 {
                     CustomMessageBox.Show(Loc.Localize("GamePathSafeguardError", "Please do not select the \"game\" or \"boot\" folder of your FFXIV installation, and choose the folder that contains these instead."), "Error",
                         MessageBoxButton.OK, MessageBoxImage.Error, parentWindow: this);
                     return;
                 }
 
-                if (!Util.IsValidFfxivPath(GamePathEntry.Text))
+                if (!GameHelpers.IsValidFfxivPath(GamePathEntry.Text))
                 {
                     if (CustomMessageBox.Show(Loc.Localize("GamePathInvalidConfirm", "The folder you selected has no FFXIV installation.\nXIVLauncher will install FFXIV the first time you log in.\nContinue?"), "XIVLauncher",
-                        MessageBoxButton.YesNo, MessageBoxImage.Information, parentWindow: this) != MessageBoxResult.Yes)
+                            MessageBoxButton.YesNo, MessageBoxImage.Information, parentWindow: this) != MessageBoxResult.Yes)
                     {
                         return;
                     }
                 }
 
-                if (Util.CanFfxivMightNotBeInternationalClient(GamePathEntry.Text))
+                if (GameHelpers.CanFfxivMightNotBeInternationalClient(GamePathEntry.Text))
                 {
                     if (CustomMessageBox.Show(Loc.Localize("GamePathRegionConfirm", "The folder you selected might be the Chinese or Korean release of the game. XIVLauncher only supports international release of the game.\nIs the folder you've selected indeed for the international version?"), "XIVLauncher",
-                        MessageBoxButton.YesNo, MessageBoxImage.Warning, parentWindow: this) != MessageBoxResult.Yes)
+                            MessageBoxButton.YesNo, MessageBoxImage.Warning, parentWindow: this) != MessageBoxResult.Yes)
                     {
                         return;
                     }
