+++ src/XIVLauncher/Windows/SettingsControl.xaml.cs
@@ -17,6 +17,7 @@ using XIVLauncher.Common.Addon;
 using XIVLauncher.Common.Addon.Implementations;
 using XIVLauncher.Common.Dalamud;
 using XIVLauncher.Common.Game.Patch.Acquisition;
+using XIVLauncher.Common.Util;
 using XIVLauncher.Support;
 using XIVLauncher.Windows.ViewModel;
 
@@ -175,7 +176,7 @@ namespace XIVLauncher.Windows
                                           .WithParentWindow(Window.GetWindow(this))
                                           .Show() == MessageBoxResult.Yes;
 
-            Util.StartOfficialLauncher(App.Settings.GamePath, isSteam, App.Settings.IsFt.GetValueOrDefault(false));
+            GameHelpers.StartOfficialLauncher(App.Settings.GamePath, isSteam, App.Settings.IsFt.GetValueOrDefault(false));
         }
 
         // All of the list handling is very dirty - but i guess it works
@@ -331,12 +332,14 @@ namespace XIVLauncher.Windows
         {
             try
             {
-                if (!string.IsNullOrEmpty(ViewModel.GamePath) && Util.IsValidFfxivPath(ViewModel.GamePath) && !DalamudLauncher.CanRunDalamud(new DirectoryInfo(ViewModel.GamePath)))
+                if (!string.IsNullOrEmpty(ViewModel.GamePath) && GameHelpers.IsValidFfxivPath(ViewModel.GamePath) && !DalamudLauncher.CanRunDalamud(new DirectoryInfo(ViewModel.GamePath)))
+                {
                     CustomMessageBox.Show(
                         Loc.Localize("DalamudIncompatible", "Dalamud was not yet updated for your current FFXIV version.\nThis is common after patches, so please be patient or ask on the Discord for a status update!"),
                         "XIVLauncher", MessageBoxButton.OK, MessageBoxImage.Asterisk, parentWindow: Window.GetWindow(this));
+                }
             }
-            catch(Exception exc)
+            catch (Exception exc)
             {
                 CustomMessageBox.Show(Loc.Localize("DalamudCompatCheckFailed",
                     "Could not contact the server to get the current compatible FFXIV version Dalamud. This might mean that your .NET installation is too old.\nPlease check the Discord for more information."), "XIVLauncher Problem", MessageBoxButton.OK, MessageBoxImage.Hand, parentWindow: Window.GetWindow(this));
@@ -350,9 +353,11 @@ namespace XIVLauncher.Windows
             string selectedPath = null;
 
             var definitionFiles = Directory.GetFiles(Path.Combine(Paths.RoamingPath, "installedPlugins"), "*.json", SearchOption.AllDirectories);
+
             foreach (var path in definitionFiles)
             {
                 dynamic definition = JObject.Parse(File.ReadAllText(path));
+
                 try
                 {
                     if (PluginListView.SelectedValue.ToString().Contains(definition.Name.Value + " " + definition.AssemblyVersion.Value))
@@ -400,7 +405,7 @@ namespace XIVLauncher.Windows
 
         private void DeletePlugin_OnClick(object sender, RoutedEventArgs e)
         {
-            if(Util.CheckIsGameOpen())
+            if (GameHelpers.CheckIsGameOpen())
             {
                 CustomMessageBox.Show(Loc.Localize("GameIsOpenPluginLocked", "The game is open, please close it and try again."), "XIVLauncher Problem", parentWindow: Window.GetWindow(this));
                 return;
@@ -511,10 +516,11 @@ namespace XIVLauncher.Windows
         {
             var isBootOrGame = false;
             var mightBeNonInternationalVersion = false;
+
             try
             {
-                isBootOrGame = !Util.LetChoosePath(ViewModel.GamePath);
-                mightBeNonInternationalVersion = Util.CanFfxivMightNotBeInternationalClient(ViewModel.GamePath);
+                isBootOrGame = !GameHelpers.LetChoosePath(ViewModel.GamePath);
+                mightBeNonInternationalVersion = GameHelpers.CanFfxivMightNotBeInternationalClient(ViewModel.GamePath);
             }
             catch (Exception ex)
             {
